Psychotip_backend/requirements.txt
bcrypt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
Markdown==3.0.1
oauthlib==3.0.1
Pillow
psycopg2==2.7.7
python3-openid==3.1.0
pylint
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
bcrypt
passlib
